refactor: move UiStates enum to dedicated state.ts file and update imports
