+++ src/components/Auth/PasskeyGeneration.tsx
@@ -3,7 +3,8 @@ import React, { type FC, useState } from 'react';
 import { Header } from '../Shared/Header';
 import { PrimaryButton } from '../Shared/PrimaryButton';
 import { useAuthContext } from '../../context/AuthContext';
-import { UiStates, useUIManager } from '../../context/UIManagerContext';
+import { UiStates } from '../../enums';
+import { useUIManager } from '../../context/UIManagerContext';
 import { Benefits } from '../Passkey/PasskeyBenefits';
 import { ErrorMessage, Loader } from '../Shared';
 import { createAccount, createPasskey } from '../../services';
