+++ src/components/Auth/PasskeyLoginFail.tsx
@@ -2,7 +2,8 @@ import { Header, LegalNotice, PrimaryButton, SecondaryButton } from '../Shared';
 import { getSignInTitle } from '../../utils/uiUtils';
 import React from 'react';
 import { useDevCredentials } from '../../context/DevCredentialsContext';
-import { UiStates, useUIManager } from '../../context/UIManagerContext';
+import { UiStates } from '../../enums';
+import { useUIManager } from '../../context/UIManagerContext';
 import { getAppUrl } from '../../utils/urlHelpers';
 import { clearLoggedEmail } from '../../services';
 import { useAuthContext } from '../../context/AuthContext';
