+++ src/components/Vehicles/CompatibilityCheck.tsx
@@ -1,7 +1,8 @@
 import { useEffect, useState, type FC } from 'react';
 
 import { PrimaryButton, ConnectedLoader, Card, Header } from '../Shared';
-import { UiStates, useUIManager } from '../../context/UIManagerContext';
+import { UiStates } from '../../enums';
+import { useUIManager } from '../../context/UIManagerContext';
 import { getAppUrl } from '../../utils/urlHelpers';
 import { AuthProvider } from '../../utils/authUrls';
 import {
