+++ src/components/Vehicles/SelectVehicles.tsx
@@ -18,7 +18,8 @@ import { useDevCredentials } from '../../context/DevCredentialsContext';
 import { getPermsValue } from '../../services/permissionsService';
 import PrimaryButton from '../Shared/PrimaryButton';
 import { backToThirdParty } from '../../utils/messageHandler';
-import { UiStates, useUIManager } from '../../context/UIManagerContext';
+import { UiStates } from '../../enums';
+import { useUIManager } from '../../context/UIManagerContext';
 import { ConnectedLoader } from '../Shared/Loader';
 import { EmptyState } from './EmptyState';
 import { ConnectCarButton } from '../Shared/ConnectCarButton';
