+++ src/context/AuthContext.tsx
@@ -28,7 +28,8 @@ import {
 } from '../utils/authUtils';
 import { useDevCredentials } from './DevCredentialsContext';
 import { UserObject } from '../models/user';
-import { UiStates, useUIManager } from './UIManagerContext';
+import { UiStates } from '../enums';
+import { useUIManager } from './UIManagerContext';
 import { TStamper } from '@turnkey/http/dist/base';
 import {
   getApiKeyStamper,
